infra/cdk/requirements.txt
aws-cdk-lib>=2.136.0
constructs>=10.3.0
boto3>=1.34.0
sagemaker>=2.214.1
